+++ ask-sdk-model/ask_sdk_model/er/dynamic/__init__.py
@@ -14,7 +14,7 @@
 #
 from __future__ import absolute_import
 
-from .entity_value_and_synonyms import EntityValueAndSynonyms
 from .update_behavior import UpdateBehavior
+from .entity_value_and_synonyms import EntityValueAndSynonyms
 from .entity_list_item import EntityListItem
 from .entity import Entity