+++ ask-sdk-model/ask_sdk_model/interfaces/amazonpay/model/request/__init__.py
@@ -14,13 +14,13 @@
 #
 from __future__ import absolute_import
 
-from .payment_action import PaymentAction
-from .price import Price
-from .seller_order_attributes import SellerOrderAttributes
-from .base_amazon_pay_entity import BaseAmazonPayEntity
-from .authorize_attributes import AuthorizeAttributes
-from .provider_attributes import ProviderAttributes
-from .billing_agreement_attributes import BillingAgreementAttributes
-from .provider_credit import ProviderCredit
-from .seller_billing_agreement_attributes import SellerBillingAgreementAttributes
 from .billing_agreement_type import BillingAgreementType
+from .seller_order_attributes import SellerOrderAttributes
+from .provider_credit import ProviderCredit
+from .authorize_attributes import AuthorizeAttributes
+from .seller_billing_agreement_attributes import SellerBillingAgreementAttributes
+from .billing_agreement_attributes import BillingAgreementAttributes
+from .base_amazon_pay_entity import BaseAmazonPayEntity
+from .payment_action import PaymentAction
+from .provider_attributes import ProviderAttributes
+from .price import Price