+++ ask-sdk-model/ask_sdk_model/interfaces/connections/requests/__init__.py
@@ -14,9 +14,9 @@
 #
 from __future__ import absolute_import
 
-from .print_pdf_request import PrintPDFRequest
-from .schedule_food_establishment_reservation_request import ScheduleFoodEstablishmentReservationRequest
 from .print_image_request import PrintImageRequest
+from .schedule_food_establishment_reservation_request import ScheduleFoodEstablishmentReservationRequest
 from .schedule_taxi_reservation_request import ScheduleTaxiReservationRequest
-from .base_request import BaseRequest
+from .print_pdf_request import PrintPDFRequest
 from .print_web_page_request import PrintWebPageRequest
+from .base_request import BaseRequest