+++ ask-sdk-model/ask_sdk_model/services/__init__.py
@@ -14,15 +14,15 @@
 #
 from __future__ import absolute_import
 
-from .base_service_client import BaseServiceClient
-from .api_response import ApiResponse
-from .api_client_response import ApiClientResponse
-from .service_client_factory import ServiceClientFactory
-from .api_client import ApiClient
-from .serializer import Serializer
-from .api_configuration import ApiConfiguration
-from .service_client_response import ServiceClientResponse
-from .service_exception import ServiceException
 from .api_client_request import ApiClientRequest
+from .api_client_response import ApiClientResponse
+from .base_service_client import BaseServiceClient
 from .api_client_message import ApiClientMessage
 from .authentication_configuration import AuthenticationConfiguration
+from .service_client_response import ServiceClientResponse
+from .api_response import ApiResponse
+from .api_configuration import ApiConfiguration
+from .serializer import Serializer
+from .service_client_factory import ServiceClientFactory
+from .api_client import ApiClient
+from .service_exception import ServiceException